fix(streamers): Don't miss the last BUFFER_SIZE records
When running `read_u32_into` we where trying to fill the whole
`click_buffer` so if the last chunk of records wasn't enough to fill it
errored and stopped the iteration.

The solution was a simple as not trying to fill in the whole buffer by
selecting only a slice big enough to fit the remaining number of
records.